chore(core): refactor fft backend
